Prevent data loss

+++ test/blackbox/types/Data64kbType.cpp
@@ -55,9 +55,9 @@ bool Data64kbType::deserialize(SerializedPayload_t* payload, void* data) {
 	return true;
 }
 
-std::function<uint32_t()> Data64kbType::getSerializedSizeProvider(void *data)
+std::function<size_t()> Data64kbType::getSerializedSizeProvider(void *data)
 {
-    return [&]() -> uint32_t { return type::getCdrSerializedSize(*static_cast<Data64kb*>(data));};
+    return [&]() -> size_t { return type::getCdrSerializedSize(*static_cast<Data64kb*>(data));};
 }
 void* Data64kbType::createData() {
 	return (void*)new Data64kb();
