+++ test/blackbox/types/HelloWorldType.h
@@ -34,7 +34,7 @@ public:
 	virtual ~HelloWorldType();
 	bool serialize(void*data,SerializedPayload_t* payload);
 	bool deserialize(SerializedPayload_t* payload,void * data);
-        std::function<uint32_t()> getSerializedSizeProvider(void *data);
+        std::function<size_t()> getSerializedSizeProvider(void *data);
 	bool getKey(void*data,InstanceHandle_t* ihandle);
 	void* createData();
 	void deleteData(void* data);
